dockerfile 생성에 따른 디렉토리 구조 변경
우분투 리눅스 환경에서 go앱을 실행시키려면 크로스 컴파일 빌드를 해야 하는데 도커나 xgo패키지의 도움 없이는 윈도우 환경에서 리눅스에 맞게 빌드를 할 수가 없다.
리눅스 환경으로 빌드를 하기 위해 dockerfile을 추가 하였고
오라클db를 사용하게 도와주는 godror패키지를 이용하기 위해선 oci(oracle client instant)와 c컴파일을 위한 gcc가 필요하여 이 패키지에 의존적일 수 밖에 없기에 이에 따른 dockerfile 이미지를 만들었고
go와 의존성패키지를 같은 도커파일로 사용하게 되면 코드 수정마다 빌드시간이 오래 걸리기 때문에
base-image와 csm-api로 디렉토리를 나누고 각각 이미지를 생성하는 dockerfile을 작성하였다.
