+++ .gitignore
@@ -1,2 +1,3 @@
 .idea/*
-config/*
+csm-api/config/*
+*.tar
