+++ csm-api/handler/response.go
@@ -1,0 +1,37 @@
+package handler
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+type ErrResponse struct {
+	Result  string   `json:"result"`
+	Message string   `json:"message"`
+	Details []string `json:"details, empty"`
+}
+
+func RespondJSON(ctx context.Context, w http.ResponseWriter, body any, status int) {
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	bodyBytes, err := json.Marshal(body)
+
+	if err != nil {
+		fmt.Printf("encode response error: %v", err)
+		w.WriteHeader(http.StatusInternalServerError)
+		rsp := ErrResponse{
+			Result:  "encode error",
+			Message: http.StatusText(http.StatusInternalServerError),
+		}
+		if err := json.NewEncoder(w).Encode(rsp); err != nil {
+			fmt.Printf("write error response error: %v", err)
+		}
+		return
+	}
+
+	w.WriteHeader(status)
+	if _, err := fmt.Fprintf(w, "%s", bodyBytes); err != nil {
+		fmt.Printf("write response error: %v", err)
+	}
+}
