+++ csm-api/service/service.go
@@ -1,0 +1,7 @@
+package service
+
+import "context"
+
+type GetUserValidService interface {
+	GetUserValid(ctx context.Context, userId string, userPwd string) (string, error)
+}
