--- server.go
@@ -1,40 +1,0 @@
-package main
-
-import (
-	"context"
-	"golang.org/x/sync/errgroup"
-	"log"
-	"net"
-	"net/http"
-)
-
-// http 서버를 실행하기 위한 struct
-type Server struct {
-	srv *http.Server
-	l   net.Listener
-}
-
-// Server struct에 담긴 값들을 바탕으로 실제 http서버를 고루틴을 이용하여 실행
-func (s *Server) Run(ctx context.Context) error {
-	eg, ctx := errgroup.WithContext(ctx)
-	eg.Go(func() error {
-		if err := s.srv.Serve(s.l); err != nil && err != http.ErrServerClosed {
-			log.Printf("failed to close: %+v", err)
-			return err
-		}
-		return nil
-	})
-	<-ctx.Done()
-	if err := s.srv.Shutdown(context.Background()); err != nil {
-		log.Printf("failed to shutdown: %+v", err)
-	}
-	return eg.Wait()
-}
-
-// Server struct 생성
-func NewServer(l net.Listener, mux http.Handler) *Server {
-	return &Server{
-		srv: &http.Server{Handler: mux},
-		l:   l,
-	}
-}
