--- store/store.go
@@ -1,10 +1,0 @@
-package store
-
-import (
-	"context"
-	"csm-api/entity"
-)
-
-type GetUserValidStore interface {
-	GetUserValid(ctx context.Context, db Queryer, userId string, userPwd string) (entity.User, error)
-}
